+++ examples/PauseOnHover.js
@@ -1,0 +1,29 @@
+import React, { Component } from 'react'
+import Slider from '../src/slider'
+
+export default class PauseOnHover extends Component {
+  render() {
+    var settings = {
+      dots: true,
+      infinite: true,
+      slidesToShow: 3,
+      slidesToScroll: 1,
+      autoplay: true,
+      autoplaySpeed: 2000,
+      pauseOnHover: true
+    };
+    return (
+      <div>
+        <h2>Pause On Hover</h2>
+        <Slider {...settings}>
+          <div><h3>1</h3></div>
+          <div><h3>2</h3></div>
+          <div><h3>3</h3></div>
+          <div><h3>4</h3></div>
+          <div><h3>5</h3></div>
+          <div><h3>6</h3></div>
+        </Slider>
+      </div>
+    );
+  }
+}
