+++ examples/UnevenSets.js
@@ -1,0 +1,27 @@
+import React, { Component } from 'react'
+import Slider from '../src/slider'
+
+export default class UnevenSets extends Component {
+  render() {
+    var settings = {
+      dots: true,
+      infinite: true,
+      speed: 500,
+      slidesToScroll: 4,
+      slidesToShow: 4
+    };
+    return (
+      <div>
+        <h2>Uneven sets</h2>
+        <Slider {...settings}>
+          <div><h3>1</h3></div>
+          <div><h3>2</h3></div>
+          <div><h3>3</h3></div>
+          <div><h3>4</h3></div>
+          <div><h3>5</h3></div>
+          <div><h3>6</h3></div>
+        </Slider>
+      </div>
+    );
+  }
+}
