+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@condenast/cn-react-slick",
-  "version": "0.11.1",
+  "version": "0.12.4",
   "description": " React port of slick carousel",
   "main": "./lib",
   "scripts": {
@@ -70,9 +70,7 @@
   "dependencies": {
     "classnames": "^2.2.5",
     "json2mq": "^0.2.0",
-    "lodash": "^3.10.1",
     "object-assign": "^4.1.0",
-    "phantomjs-prebuilt": "^2.1.7",
     "react-responsive-mixin": "^0.3.5",
     "slick-carousel": "^1.6.0"
   },
