+++ src/arrows.jsx
@@ -6,7 +6,7 @@ import classnames from 'classnames';
 export var PrevArrow = React.createClass({
 
   clickHandler: function (options, e) {
-    e.preventDefault();
+    if (e) { e.preventDefault(); }
     this.props.clickHandler(options, e);
   },
   render: function () {
@@ -40,7 +40,7 @@ export var PrevArrow = React.createClass({
 
 export var NextArrow = React.createClass({
   clickHandler: function (options, e) {
-    e.preventDefault();
+    if (e) { e.preventDefault(); }
     this.props.clickHandler(options, e);
   },
   render: function () {
