+++ src/dots.jsx
@@ -30,8 +30,10 @@ export var Dots = React.createClass({
     // Credit: http://stackoverflow.com/a/13735425/1849458
     var dots = Array.apply(null, Array(dotCount + 1).join('0').split('')).map((x, i) => {
 
+      var leftBound = (i * this.props.slidesToScroll);
+      var rightBound = (i * this.props.slidesToScroll) + (this.props.slidesToScroll - 1);
       var className = classnames({
-        'slick-active': (this.props.currentSlide === i * this.props.slidesToScroll)
+        'slick-active': (this.props.currentSlide >= leftBound) && (this.props.currentSlide <= rightBound)
       });
 
       var dotOptions = {
