+++ src/mixins/event-handlers.js
@@ -7,19 +7,22 @@ var EventHandlers = {
   // Event handler for previous and next
   changeSlide: function (options) {
     var indexOffset, previousInt, slideOffset, unevenOffset, targetSlide;
-    unevenOffset = (this.state.slideCount % this.props.slidesToScroll !== 0);
-    indexOffset = unevenOffset ? 0 : (this.state.slideCount - this.state.currentSlide) % this.props.slidesToScroll;
+    const {slidesToScroll, slidesToShow} = this.props
+    const {slideCount, currentSlide} = this.state
+    unevenOffset = (slideCount % slidesToScroll !== 0);
+    indexOffset = unevenOffset ? 0 : (slideCount - currentSlide) % slidesToScroll;
 
     if (options.message === 'previous') {
-      slideOffset = (indexOffset === 0) ? this.props.slidesToScroll : this.props.slidesToShow - indexOffset;
-      targetSlide = this.state.currentSlide - slideOffset;
+      slideOffset = (indexOffset === 0) ? slidesToScroll : slidesToShow - indexOffset;
+      targetSlide = currentSlide - slideOffset;
       if (this.props.lazyLoad) {
-        previousInt = this.state.currentSlide - slideOffset;
-        targetSlide = previousInt === -1 ? this.state.slideCount -1 : previousInt;
+        previousInt = currentSlide - slideOffset;
+        targetSlide = previousInt === -1 ? slideCount -1 : previousInt;
       }
     } else if (options.message === 'next') {
-      slideOffset = (indexOffset === 0) ? this.props.slidesToScroll : indexOffset;
-      targetSlide = this.state.currentSlide + slideOffset;
+      slideOffset = (indexOffset === 0) ? slidesToScroll : indexOffset;
+      // targetSlide = currentSlide + slideOffset;
+      targetSlide = ((currentSlide + slidesToScroll) % slideCount) + indexOffset;
     } else if (options.message === 'dots') {
       // Click on dots
       targetSlide = options.index * options.slidesToScroll;
@@ -65,6 +68,7 @@ var EventHandlers = {
   },
   swipeMove: function (e) {
     if (!this.state.dragging) {
+      e.preventDefault();
       return;
     }
     if (this.state.animating) {
@@ -120,6 +124,7 @@ var EventHandlers = {
   },
   swipeEnd: function (e) {
     if (!this.state.dragging) {
+      e.preventDefault();
       return;
     }
     var touchObject = this.state.touchObject;
